raceresults/requirements.txt
Django==1.9.5
wheel==0.24.0
psycopg2==2.6.1

django-allauth==0.25.2

opbeat==3.3.2
django-autofixture==0.12.1
